# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tasing API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Tasing API Server"
        }
    ],
    "paths": {
...
        "/api/email/verify": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get the email verification status (true or false)",
                "operationId": "show",
                "responses": {
                    "200": {
                        "description": "Verification status returned.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "properties": {
                                                "has_verified_email": {
                                                    "type": "boolean"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    }
                },
                "security": [
                    {
                        "": []
                    }
                ]
            }
        },
        "/api/email/verify/{id}/{hash}": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Mark the authenticated user's email address as verified",
                "operationId": "verify",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "hash",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "expires",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "signature",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Verified successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "202": {
                        "description": "Already verified."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Invalid signature; or user logged isn't the same as user looking for verification."
                    }
                },
                "security": [
                    {
                        "": []
                    }
                ]
            }
        },
        "/api/email/resend": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Resend the email verification notification",
                "operationId": "resend",
                "responses": {
                    "200": {
                        "description": "Verified successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "202": {
                        "description": "Already verified."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    }
                },
                "security": [
                    {
                        "": []
                    }
                ]
            }
        },
...
    },
...
}